Bug fix: uid type is long.
git-svn-id: https://svn.erp5.org/repos/public/erp5/trunk@3777 20353a03-c40f-0410-a6d1-a30d3c3de9de

+++ product/ERP5Type/PropertySheet/Base.py
@@ -40,6 +40,6 @@ class Base:
             'mode'        : '' },
         {   'id'          : 'uid',
             'description' : 'Unique ID of the object in the ZSQLCatalog',
-            'type'        : 'int',
+            'type'        : 'long',
             'mode'        : '' },
     )
